requirements.txt
Flask==2.3.3
flask-socketio==5.3.4
python-socketio==5.9.0
eventlet==0.33.3
qrcode==7.4.2
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0